all: Fix error 'truncation is not allowed in initializer list' for CLang

+++ src/game/client/c_baseentity.cpp
@@ -3753,7 +3753,7 @@ void C_BaseEntity::AddColoredDecal( const Vector& rayStart, const Vector& rayEnd
 
 	case mod_brush:
 		{
-			color32 cColor32 = { cColor.r(), cColor.g(), cColor.b(), cColor.a() };
+			color32 cColor32 = { (byte)cColor.r(), (byte)cColor.g(), (byte)cColor.b(), (byte)cColor.a() };
 			effects->DecalColorShoot( decalIndex, index, model, GetAbsOrigin(), GetAbsAngles(), decalCenter, 0, 0, cColor32 );
 		}
 		break;
